requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.1
networkx==2.2
numpy==1.15.4
Pillow==5.3.0
pony==0.7.6
praw==6.0.0
prawcore==1.0.0
psycopg2==2.7.6
pyparsing==2.3.0
pytesseract==0.2.5
python-dateutil==2.7.5
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.28.1
update-checker==0.16
urllib3==1.24.1
Werkzeug==0.14.1
